fix(Directory organization) The structure has been organized and standardized
