+++ tools/bedrock/http/decompact_file.go
